Longest Lines
 Write a program which reads a file and prints to stdout the specified number of the longest lines that are sorted based on their length in descending order.

Your program should accept a path to a file as its first argument. The file contains multiple lines. The first line indicates the number of lines you should output, the other lines are of different length and are presented randomly. You may assume that the input file is formatted correctly and the number in the first line is a valid positive integer.

Print out the longest lines limited by specified number and sorted by their length in descending order.