Remove Characters
Write a program which removes specific characters from a string. 

The first argument is a path to a file. The file contains the source strings and the characters that need to be scrubbed. Each source string and characters you need to scrub are delimited by comma. 

Print to stdout the scrubbed strings, one per line. Ensure that there are no trailing empty spaces on each line you print.